+++ webknossos/webknossos/dataset/layer/__init__.py
@@ -22,6 +22,7 @@ from .view import (
     ArrayInfo,
     BaseArray,
     MagView,
+    RemoteMagView,
     TensorStoreArray,
     View,
     Zarr3ArrayInfo,